refactor(state): split registered agent database listing from the registry
