+++ src/state/openclaw-agent-db-registry-listing.ts
@@ -1,0 +1,132 @@
+import { existsSync, lstatSync, statSync } from "node:fs";
+import path from "node:path";
+import { executeSqliteQuerySync, getNodeSqliteKysely } from "../infra/kysely-sync.js";
+import { resolveSqliteDatabaseFilePaths } from "../infra/sqlite-files.js";
+import { normalizeAgentId } from "../routing/session-key.js";
+import type { OpenClawRegisteredAgentDatabase } from "./openclaw-agent-db-contract.js";
+import { withOpenClawStateDatabaseReadOnly } from "./openclaw-state-db-readonly.js";
+import { detectOpenClawStateDatabaseSchemaMigrationsFromDatabase } from "./openclaw-state-db-schema-repair.js";
+import type { DB as OpenClawStateKyselyDatabase } from "./openclaw-state-db.generated.js";
+import type { OpenClawStateDatabaseOptions } from "./openclaw-state-db.js";
+import { resolveOpenClawStateSqlitePath } from "./openclaw-state-db.paths.js";
+
+type OpenClawAgentRegistryDatabase = Pick<OpenClawStateKyselyDatabase, "agent_databases">;
+
+// Registry metadata is process-stable: registry writes invalidate after each commit;
+// other-process changes take effect on restart. Polling here puts schema probes back on hot reads.
+let registeredAgentDatabasesMemo:
+  | {
+      pathname: string;
+      entries: readonly OpenClawRegisteredAgentDatabase[];
+    }
+  | undefined;
+
+function resolveAgentDatabaseRegistryPath(options: OpenClawStateDatabaseOptions): string {
+  return path.resolve(options.path ?? resolveOpenClawStateSqlitePath(options.env ?? process.env));
+}
+
+export function invalidateRegisteredAgentDatabasesMemo(
+  options: OpenClawStateDatabaseOptions,
+): void {
+  const pathname = resolveAgentDatabaseRegistryPath(options);
+  if (registeredAgentDatabasesMemo?.pathname === pathname) {
+    registeredAgentDatabasesMemo = undefined;
+  }
+}
+
+function cloneRegisteredAgentDatabases(
+  entries: readonly OpenClawRegisteredAgentDatabase[],
+): OpenClawRegisteredAgentDatabase[] {
+  return entries.map((entry) => ({ ...entry }));
+}
+
+function hasUnavailableMissingSqlitePath(pathname: string): boolean {
+  for (const candidate of resolveSqliteDatabaseFilePaths(pathname)) {
+    try {
+      lstatSync(candidate);
+      return true;
+    } catch (error) {
+      if ((error as NodeJS.ErrnoException).code !== "ENOENT") {
+        return true;
+      }
+    }
+  }
+
+  let ancestor = path.dirname(pathname);
+  while (true) {
+    try {
+      const stat = lstatSync(ancestor);
+      if (!stat.isSymbolicLink()) {
+        return !stat.isDirectory();
+      }
+      try {
+        return !statSync(ancestor).isDirectory();
+      } catch {
+        return true;
+      }
+    } catch (error) {
+      if ((error as NodeJS.ErrnoException).code !== "ENOENT") {
+        return true;
+      }
+    }
+    const parent = path.dirname(ancestor);
+    if (parent === ancestor) {
+      return false;
+    }
+    ancestor = parent;
+  }
+}
+
+/** List agent databases recorded in the shared OpenClaw state registry. */
+export function listOpenClawRegisteredAgentDatabases(
+  options: OpenClawStateDatabaseOptions = {},
+): OpenClawRegisteredAgentDatabase[] {
+  const pathname = resolveAgentDatabaseRegistryPath(options);
+  if (registeredAgentDatabasesMemo?.pathname === pathname) {
+    return cloneRegisteredAgentDatabases(registeredAgentDatabasesMemo.entries);
+  }
+  if (!existsSync(pathname)) {
+    if (hasUnavailableMissingSqlitePath(pathname)) {
+      throw new Error(`OpenClaw state database ${pathname} is unavailable.`);
+    }
+    registeredAgentDatabasesMemo = { pathname, entries: [] };
+    return [];
+  }
+  // Discovery runs per row in list hot paths, so the legacy-schema gate and the
+  // query share one process-held state handle instead of opening two
+  // connections per call.
+  const entries = withOpenClawStateDatabaseReadOnly(({ db: database }) => {
+    if (detectOpenClawStateDatabaseSchemaMigrationsFromDatabase(database, pathname).length > 0) {
+      throw new Error(
+        `OpenClaw state database ${pathname} has a legacy agent database registry schema; run openclaw doctor --fix to migrate it.`,
+      );
+    }
+    const registryTable = database
+      .prepare("SELECT type FROM sqlite_master WHERE name = 'agent_databases'")
+      .get() as { type?: unknown } | undefined;
+    if (!registryTable) {
+      return [];
+    }
+    if (registryTable.type !== "table") {
+      throw new Error(`OpenClaw state database ${pathname} has an invalid agent registry.`);
+    }
+    const db = getNodeSqliteKysely<OpenClawAgentRegistryDatabase>(database);
+    const rows = executeSqliteQuerySync(
+      database,
+      db
+        .selectFrom("agent_databases")
+        .selectAll()
+        .orderBy("agent_id", "asc")
+        .orderBy("path", "asc"),
+    ).rows;
+    return rows.map((row) => ({
+      agentId: normalizeAgentId(row.agent_id),
+      path: row.path,
+      schemaVersion: row.schema_version,
+      lastSeenAt: row.last_seen_at,
+      sizeBytes: row.size_bytes,
+    }));
+  }, options);
+  registeredAgentDatabasesMemo = { pathname, entries };
+  return cloneRegisteredAgentDatabases(entries);
+}
